requirements.txt
discord.py
python-dotenv
pytz
schedule